CDK_Code/requirements.txt
aws-cdk-lib==2.147.3
constructs>=10.0.0,<11.0.0
python-dotenv